+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 Change log for the Basil Game Engine project.
 
+## 2022-08-03
+### Added
+* Script Component UI to Pesto.
+
+
 ## 2022-08-02
 ### Added
 * Getting an entity by its UUID.